Reverting to old commit before using Gatsby Script component
