+++ gatsby-browser.js
@@ -1,7 +1,29 @@
 import React from 'react';
 import SkipLinks from './src/components/skip-links';
 
-export { wrapPageElement } from './gatsby-shared';
+export const wrapPageElement = ({ element }) => {
+  // Props provide same data to Layout as Page element will get
+  // Including location, data, etc - you don't need to pass it
+  return (
+    <React.Fragment>
+      <div
+        css={{
+          minHeight: '100%',
+          display: 'grid',
+          gridTemplateRows: 'auto auto 1fr',
+          gridTemplateColumns: '100%'
+        }}
+      >
+        <div>
+          <SkipLinks />
+          <m-universal-header></m-universal-header>
+        </div>
+        {element}
+      </div>
+      <m-chat id='chat'></m-chat>
+    </React.Fragment>
+  );
+};
 
 export const onRouteUpdate = ({ location, prevLocation }) => {
   const newPath = location.pathname;
